requirement.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
psycopg2-binary==2.9.7